refactor: reorganize packaging scripts and update paths for macOS distribution assembly
