+++ tools/packaging/__init__.py
@@ -1,0 +1,1 @@
+"""CI 与 Nuitka 发行组装脚本。"""
