+++ tools/packaging/copy_maagentbinary_to_dist.py
