+++ tools/packaging/fix_maafw_macos_dylib_paths.py
